+++ src/app/app.component.ts
@@ -1,4 +1,3 @@
-import { HomePage } from './../pages/home/home'
 import { Component, ViewChild } from '@angular/core'
 
 import { Platform, MenuController, Nav } from 'ionic-angular'
@@ -7,6 +6,7 @@ import { StatusBar, Splashscreen } from 'ionic-native'
 
 import { HelloIonicPage } from '../pages/hello-ionic/hello-ionic'
 import { ListPage } from '../pages/list/list'
+import { LoginPage } from '../pages/login/login'
 
 
 @Component({
@@ -16,7 +16,7 @@ export class MyApp {
   @ViewChild(Nav) nav: Nav
 
   // make HelloIonicPage the root (or first) page
-  rootPage: any = HomePage
+  rootPage: any = LoginPage
   pages: Array<{title: string, component: any}>
 
   constructor(